add billing plan config
